OnlineCourses/requirements.txt
Django~=3.2.3
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
mysqlclient==2.0.3